deps: fix deps so that py3.6 actually works
Before, I couldn't run `poetry update` because
there would be dependency conflicts, since
pyarrow 7.0.0 requires python >3.7.

In addition, there was a bug
https://github.com/pytest-dev/pytest/issues/10420
in pytest-benchmark <4.0 that was giving me errors when I had python>3.6
so I upgraded that when
we have python >3.7.
But pytest-benchmark 4 isn't compatible with python3.6, so I had
to split the deps into two different cases.

I tweaked the min python version from 3.6 to 3.6.2
because some packages were incompatible with
3.6.0 to 3.6.2, so it was easier to
require just a slightly higher python

Note that although now poetry is happy to
solve dependencies for python 3.6,
python 3.6 still doesn't actually work
due to the `from __future__ import annotations`
in linker.py. But this has a been a problem for months and we
just seem to be ignoring the
[failing 3.6 tests in CI](https://github.com/moj-analytical-services/splink/actions/workflows/pytest_36_compatibility.yml)

Whatever though, at least this unblocks me so now I can actually develop
splink. I think the next step is to actually remove 3.6
support, since 3.6 reached
end of life in Dec 2021. Separate PR for that though.