feat: Support sqlglot versions >=5.1.0
Fixes
https://github.com/moj-analytical-services/splink/issues/991

Note of caution:
poetry.lock has locked the sqlglot version to
10.5.10, so in the future that is the only version
that we will be testing in a dev environment and in CI.
I did check that 5.1.0 still works by
`poetry shell && pip install sqlglot==5.1.0 && pytest tests`,
but we might unwittingly make some changes in the future that
break 5.1.0 compatibilty and we won't know.

One solution would be to actually test 5.1.0 in CI,
but I don't think that effort is worth it. If someone
files a bug report then we fix it, but otherwise
I think we can trust that the sqlglot API is stable enough
that the odds of this are low. Plus I think we can update our
requirements to sqlglot>=6.0.0 in ~6-12 months, so we won't be
living with this risk for a huge long while.