dev: Replace flake8 with ruff
It is MUCH faster (making it reasonable
to use as a pre-commit hook),
and supports isort and bugbear
so we can catch more problems,
and it supports autofixing.

I'll enable bugbear and isort in a followup, but you can see here
how much of a drop it replacement it is.