Loosen dependency ranges
This is simlar to https://github.com/moj-analytical-services/splink/issues/991

When we specify `duckdb = "^0.6.0"`, this
completely prevents me from installing 0.7.0, which was
recently released but works just fine with splink.

Generally, as a library we want to specify the loosest possible ranges.
Usually, it's the application developer
(or someone who has a full picture of
the dependency tree) who wants to use the
^ specifier.

We can keep the dev-dependencies as ^, because no downstream users are
limited by this choice.