ref: Share SQL logging logic, show SQL on errors
Before, if the backend had a problem when executing SQL,
you couldn't actually see the SQL that caused it unless
you had logging enabled.

This also removes the duplicated logging calls that were in
every linker subclass

This also makes it so that the logged SQL is consistent: In duckdb
and sqlite, the SQL was logged, and then further modified
by wrapping it in a "CREATE TABLE ...",
so the actual SQL ran was different
from what was logged.