Loosen dependency requirements
Libraries should specify the loosest dependencies possible. They should ONLY exclude versions that they KNOW to be broken. This gives users at the app level the most freedom to choose versions of packages.

For instance, I needed sqlglot>=20 in order to use a new version of ibis, but splink was restricting the version to <19, so pip wasn't able to find a suitable version. But sqlglot >=19 probably works just fine with splink (admittedly, I haven't tested it...)

If a user finds a version that is broken, they will report it, and/or they can restrict the version themselves in their requirements. But it is impossible for users to loosen requirements.

related: https://github.com/moj-analytical-services/splink/pull/1642